refactor: make image option view more generic
